requirements.txt
python-snap7
Flask
numpy
pyodbc